feat(parse): declare items and move tests
